+++ apps/ledger-live-mobile/src/newArch/features/LandingPages/screens/LargeMoverLandingPage/types.d.ts
@@ -1,16 +1,16 @@
 type InformationsProps = {
   price: number;
   ticker: string;
-  marketCap: number;
+  marketCap: number | undefined;
   volume: number;
   marketCapPercent: number;
   fdv: number;
   circulatingSupply: number;
   totalSupply: number;
   allTimeHigh: number;
-  allTimeHighDate: string;
+  allTimeHighDate: Date;
   allTimeLow: number;
-  allTimeLowDate: string;
+  allTimeLowDate: Date;
 };
 
 export type { InformationsProps };