+++ apps/ledger-live-mobile/src/screens/Account/ListHeaderComponent.tsx
@@ -36,7 +36,6 @@ import WarningBannerStatus from "~/components/WarningBannerStatus";
 import ErrorWarning from "./ErrorWarning";
 import { useFeature } from "@ledgerhq/live-common/featureFlags/index";
 import { isNFTCollectionsDisplayable } from "./nftHelper";
-import { getCurrencyConfiguration } from "@ledgerhq/live-common/config/index";
 
 type Props = {
   account?: AccountLike;
@@ -89,6 +88,7 @@ export function useListHeaderComponents({
   listHeaderComponents: ReactNode[];
   stickyHeaderIndices?: number[];
 } {
+  const llmNftSupport = useFeature("llNftSupport");
   const llmSolanaNfts = useFeature("llmSolanaNfts");
   if (!account) return { listHeaderComponents: [], stickyHeaderIndices: undefined };
 
@@ -135,12 +135,14 @@ export function useListHeaderComponents({
     isStuckOperation({ family: mainAccount.currency.family, operation: oldestEditableOperation });
 
   const displayNftCollections = isNFTCollectionsDisplayable(account, empty, {
-    llmSolanaNftsEnabled: llmSolanaNfts?.enabled,
+    llmNftSupportEnabled: !!llmNftSupport?.enabled,
+    llmSolanaNftsEnabled: !!llmSolanaNfts?.enabled,
   });
 
-  const coinConfig = getCurrencyConfiguration(currency);
   const disableDelegation =
-    "disableDelegation" in coinConfig && coinConfig.disableDelegation === true;
+    currencyConfig &&
+    "disableDelegation" in currencyConfig &&
+    currencyConfig.disableDelegation === true;
 
   return {
     listHeaderComponents: [