+++ libs/coin-modules/coin-filecoin/jest.config.js
@@ -19,4 +19,5 @@ module.exports = {
       { outputName: "filecoin-sonar-executionTests-report.xml", reportedFilePath: "absolute" },
     ],
   ],
+  setupFilesAfterEnv: ["@ledgerhq/disable-network-setup"],
 };