+++ libs/coin-modules/coin-solana/src/signer.ts
@@ -16,7 +16,19 @@ export type Resolution = {
   };
 };
 
+export enum PubKeyDisplayMode {
+  LONG,
+  SHORT,
+}
+
+export type AppConfig = {
+  blindSigningEnabled: boolean;
+  pubKeyDisplayMode: PubKeyDisplayMode;
+  version: string;
+};
+
 export interface SolanaSigner {
+  getAppConfiguration(): Promise<AppConfig>;
   getAddress(path: string, display?: boolean): Promise<SolanaAddress>;
   signTransaction(
     path: string,