+++ libs/coin-modules/coin-tezos/jest.config.js
@@ -18,4 +18,5 @@ module.exports = {
       { outputName: "tezos-sonar-executionTests-report.xml", reportedFilePath: "absolute" },
     ],
   ],
+  // setupFilesAfterEnv: ["@ledgerhq/disable-network-setup"],
 };