+++ libs/ledger-live-common/src/families/cosmos/datasets/babylon.ts
@@ -11,13 +11,11 @@ const dataset: CurrenciesData<Transaction> = {
       name: "babylon seed 1",
       apdus: `
       => 55040000180362626e2c00008076000080000000800000000000000000
-      <= 0388459b2653519948b12492f1a0b464720110c147a8155d23d423a5cc3c21d89a62626e316738343933346a70753376356465357971756b6b6b68786d6376737733753261396361736a789000
+      <= 03cd9e27aa78eddce3fda5f8050db35196b188d80d10bd04aab15e5669defee8b062626e3175777077733037376130613970636c61707633663266796a3575306d6c683665776d6473386e9000
       => 55040000180362626e2c00008076000080000000800000000000000000
-      <= 0388459b2653519948b12492f1a0b464720110c147a8155d23d423a5cc3c21d89a62626e316738343933346a70753376356465357971756b6b6b68786d6376737733753261396361736a789000
+      <= 03cd9e27aa78eddce3fda5f8050db35196b188d80d10bd04aab15e5669defee8b062626e3175777077733037376130613970636c61707633663266796a3575306d6c683665776d6473386e9000
       => 55040000180362626e2c00008076000080010000800000000000000000
-      <= 02624ac83690d5ef627927104767d679aef73d3d3c9544abe4206b1d0c463c94ff62626e31303875793571396a743539677775677135797264686b7a6364396a7279736c6d6b78703666649000
-      => 55040000180362626e2c00008076000080020000800000000000000000
-      <= 038ff98278402aa3e46ccfd020561dc9724ab63d7179ca507c8154b5257c7d520062626e3163676336393661793270673664346763656a656b3279386c6136366a37653579787830386d689000
+      <= 02b3c7368432b998240a2806eb08a00becd359867ca9cdc94f3eb4ba29a6eb0d7a62626e316a7932387a6e6c6b7034783767663863786c6a366e326e676b726b78683279793774326d39799000
       `,
     },
   ],
@@ -25,36 +23,42 @@ const dataset: CurrenciesData<Transaction> = {
     {
       FIXME_tests: ["balance is sum of ops", "pendingOperations are cleaned up"],
       raw: {
-        id: "js:2:babylon:bbn1g84934jpu3v5de5yqukkkhxmcvsw3u2a9casjx:",
-        seedIdentifier: "0388459b2653519948b12492f1a0b464720110c147a8155d23d423a5cc3c21d89a",
-        xpub: "bbn1g84934jpu3v5de5yqukkkhxmcvsw3u2a9casjx",
+        id: "js:2:babylon:bbn1uwpws077a0a9pclapv3f2fyj5u0mlh6ewmds8n:",
+        seedIdentifier: "03cd9e27aa78eddce3fda5f8050db35196b188d80d10bd04aab15e5669defee8b0",
+        used: true,
         derivationMode: "",
         index: 0,
-        freshAddress: "bbn1g84934jpu3v5de5yqukkkhxmcvsw3u2a9casjx",
+        freshAddress: "bbn1uwpws077a0a9pclapv3f2fyj5u0mlh6ewmds8n",
         freshAddressPath: "44'/118'/0'/0/0",
-        name: "Babylon 1 - Nano X",
-        blockHeight: 71848,
-        balance: "40699912",
-        spendableBalance: "40699912",
+        blockHeight: 207592,
+        creationDate: "2025-05-02T10:23:08.000Z",
+        operationsCount: 1,
         operations: [],
-        operationsCount: 2,
         pendingOperations: [],
         currencyId: "babylon",
-        lastSyncDate: "",
-        creationDate: "2025-04-10T15:47:19.000Z",
+        lastSyncDate: "2025-05-02T10:24:44.245Z",
+        balance: "1000000",
+        spendableBalance: "1000000",
+        balanceHistoryCache: {
+          HOUR: { balances: [0], latestDate: 1746180000000 },
+          DAY: { balances: [0], latestDate: 1746136800000 },
+          WEEK: { balances: [0], latestDate: 1745704800000 },
+        },
+        xpub: "bbn1uwpws077a0a9pclapv3f2fyj5u0mlh6ewmds8n",
+        swapHistory: [],
       },
       transactions: [
         {
           name: "Normal transaction",
           transaction: fromTransactionRaw({
-            amount: "1000000",
+            amount: "1000",
             recipient: "bbn1vh34djka7ug2gww9njrsmmr7emj3dx3paz5sj4",
             useAllAmount: false,
             family: "cosmos",
             mode: "send",
             networkInfo: null,
-            fees: "213",
-            gas: "106317",
+            fees: "238",
+            gas: "118993",
             validators: [],
             memo: "",
             sourceValidator: "",
@@ -62,7 +66,7 @@ const dataset: CurrenciesData<Transaction> = {
           expectedStatus: () => ({
             errors: {},
             warnings: {},
-            amount: BigNumber("1000000"),
+            amount: BigNumber("1000"),
           }),
         },
       ],